Split `sandbox_util.py` into various part of file, decrease the function stack (#38)
* Split the sandbox_util.py to break the hierarchical in the same file

* Adjust initialize and finish utility

* Let sandbox_util.py fit the utils

* Rename the sandbox_util.py to util.py

* Adjust the import section

* Rename the function since the name conflict

* Fix the function name since the function is renamed

* Remove redundant file & add Language enum to utils/sandbox/enum.py
